backend/requirements.txt
Flask==3.0.3
Flask-Cors==4.0.1
Flask-SQLAlchemy==3.1.1
Flask-Migrate==4.0.5
python-dotenv==1.0.1
SQLAlchemy==2.0.30
PyMySQL==1.1.1
psycopg2-binary==2.9.10
gunicorn==22.0.0
Flask-Login==0.6.3
passlib==1.7.4
Flask-Caching==2.1.0
pandas==2.1.4
numpy==1.24.3
scikit-learn==1.3.2
